Refined the definition of FilePermissions

+++ Kernel/Sources/klib/Types.h
@@ -163,9 +163,12 @@ typedef char                Character;
 
 // Various Kernel API types
 typedef Int     ProcessId;
+
 typedef Int32   FilesystemId;
 typedef Int32   InodeId;    // XXX should probably be 64bit
-typedef UInt32  FilePermissions;
+
+typedef UInt16  FilePermissions;
+
 typedef UInt32  UserId;
 typedef UInt32  GroupId;
